Rewrite grunt plugin on Atom
Rewrite the plugin issuing additional commands to trigger grunt.
This enables to not fork the nodejs plugin anymore, but just extends them.
